+++ src/components/AddAlbum/index.js
@@ -27,19 +27,19 @@ function AddAlbum(props){
         <div>
             <h2>Add your own!</h2>
             <form onSubmit={handleSubmit}>
-                <label HTMLfor="album">Album:</label>
+                <label htmlFor="album">Album:</label>
                 <input onChange={(event) => handleInput(event, setAlbum)} type="text" name="album" id="album" value={album}/>
 
-                <label HTMLfor="artist">Artist:</label>
+                <label htmlFor="artist">Artist:</label>
                 <input onChange={(event) => handleInput(event, setArtist)} type="text" name="artist" id="artist" value={artist}/>
 
-                <label HTMLfor="image">Cover image url:</label>
+                <label htmlFor="image">Cover image url:</label>
                 <input onChange={(event) => handleInput(event, setImage)} type="text" name="image" id="image"  value={image}/>
                 
-                <label HTMLfor="producer">Producer:</label>
+                <label htmlFor="producer">Producer:</label>
                 <input onChange={(event) => handleInput(event, setProducer)} type="text" name="producer" id="producer" value={producer}/>
 
-                <label HTMLfor="year">Year:</label>
+                <label htmlFor="year">Year:</label>
                 <input onChange={(event) => handleInput(event, setYear)} type="number" name="year" id="year" value={year}/>
 
                 <input type="submit" name="submit" id="submit" value="submit"/>
