+++ src/components/CardList/index.js
@@ -1,0 +1,11 @@
+import React, {useState} from 'react'
+
+import { Card } from '..'
+
+function CardList( { albumList }){
+    return (
+        albumList.map((album, i) => <Card key={i} data={album} />)
+    )
+}
+
+export default CardList
