+++ src/components/Main/index.js
@@ -1,18 +1,8 @@
 import React, {useState} from 'react'
 
-import Card from './Card'
-import AddAlbum from './AddAlbum'
+import {CardList, AddAlbum, Header} from '..'
 
-import album_data from '../data'
-import Header from './Header'
-
-import '../style.css'
-
-function CardList(props){
-    return (
-        props.albumList.map(album => <Card data={album} />)
-    )
-}
+import album_data from '../../data'
 
 function Main(){
