+++ src/components/index.js
@@ -1,0 +1,5 @@
+export { default as Main } from './Main';
+export { default as CardList } from './CardList';
+export { default as Header } from './Header';
+export { default as AddAlbum } from './AddAlbum';
+export { default as Card } from './Card';
